Export experiment generation to test utils
